Reduce compile warnings (#63)
* Wave 1 of compile warning reduction

* Wave 2 of compile error reduction

* Wave 3 of compile warning reduction

* Fixed heap size warning from ShaderPacker.cpp

* Fixed unsafe ctime usage within KDebug.cpp

+++ KRAWLER/KRAWLER Engine/src/AssetLoader/KAssetLoader.cpp
@@ -26,7 +26,7 @@ constexpr int32 MAX_ANIMATION_FILE_CHARS = 100000;
 std::wstring FindFilename(std::wstring& str)
 {
 	std::wstring returnStr;
-	uint32 idx = str.size() - 1;
+	uint32 idx = static_cast<unsigned>(str.size() - 1);
 	bool bIsFilename = false;
 	while (idx != 0)
 	{
